Perturbation/linearization numerics overhaul (#104)
* Drop dead D.astype dtype guards from shock-matrix solves

PyTensor's sparse_jacobian yields float64; the casts only guarded
against SymPy Integer constants from the old autodiff path.

* Harden gensys numba path: non-raising SVD, fewer copies, no object-mode

Route the SVD through pytensor's _svd_gesdd dispatch targets so a bad
draw returns NaN (rejected by the sampler) instead of raising. Defer
caching to pytensor via numba_njit (safe fastmath, no nnan/ninf). Add
infer_shape/gufunc_signature so the Op no longer falls back to object
mode under Blockwise. Destroy dead buffers in place (overwrite flags,
G_0.T LU trick), drop defensive contiguity coercion, upcast only when
the input isn't float64, and hoist the duplicated T_mat @ Q.

* Make cycle_reduction Blockwise-safe and allocation-free

Add infer_shape/gufunc_signature so the Op compiles under Blockwise
without an object-mode fallback. Rewrite the njit core to run the loop
allocation-free: preallocated C-order matmul outputs (np.dot out=),
zero-copy .T Fortran scratch for the LU/solves with overwrite=True,
in-place updates, and in-place pivot reuse. Upcast inputs only when
not float64, and alias rather than copy the rebound loop snapshots.

* Update gauss-newton CG step to update r and d in place

Drop the redundant copy on d = -r (negation already allocates) and
update the residual and direction in place across CG iterations.

* Modernize cycle_reduction numba Op and fix input-buffer overwrite

- Rename nb_cycle_reduction/nb_solve_shock_matrix to *_numpy; drop dead
  commented njit blocks.
- Derive output dtype via linalg_output_dtype instead of hardcoding
  float64; cast funcify inputs only when they differ.
- Defer njit caching/fastmath to pytensor (numba_basic.numba_njit).
- Keep kernel inputs read-only; pass overwrite_b=False to the final
  solve, which previously could clobber a Fortran-ordered input.

* Improve differentiability of BK condition

* Tighten linearize_model: single-pass incidence, hoist B/D

Derive all four lag/lead incidence arrays in one pass (one atoms walk
per equation instead of up to 2*n_vars+1). Hoist B/D/permuted_eqs out
of the use_split branch since they were identical in both arms, drop
the n_var/n_vars shadow, and use strict=True on the variable zips.

* Loosen steady-state numerical-vs-analytic test tolerance

trust-constr converges to its own gtol/xtol, so the numerical steady
state need not match the analytic one to the default rtol=1e-7.

* Replace sparse_jacobian with symbolic-diff jacobians

sparse_jacobian's graph-coloring compression was net overhead at DSGE
sizes (small, dense-consumed jacobians): its connectivity scan, coloring
and advanced-indexing assembly cost more to build and compile than they
saved. Differentiate symbolically and materialize via the sympytensor
matrix printer instead (one scatter, constants baked); fall back to plain
autodiff for the steady-state jac/hess, where no sympy form is available.
~2-3x faster end-to-end model compilation.

* Remove block-dot rewrite (local_dot_of_join)

Splitting dot(Join, Y) into per-block dots is a measured net loss on the
perturbation/cycle-reduction path: it fires on the selection-matrix dot,
turns one GEMM into smaller ones plus concat, and is ~24-40% slower to
compile and ~2-4% slower to exec. Keep the block_diag and Join utility
rewrites, which are wins.

* Share subexpressions across A/B/C/D via joint sp.cse in linearization

Build all four perturbation jacobians in one build_symbolic_jacobians call
so sp.cse pre-shares common terms across their entries before pt.grad
replicates the linearization graph. Cuts the fused logp_dlogp compile by
~12.5% on a 44-variable/46-prior model; neutral on small models.

* Tune CSE settings for best compile speed

* Solve bounded steady states by unconstrained reparametrization

A positivity bound flipped the minimize solver to trust-constr, whose
interior-point path stalls on sub-1e-12 FP noise on native x86 (red CI).
Reparametrize the objective via PyMC transforms and solve unconstrained
instead of switching methods; add a prefer_transform flag.

* Add unit_interval assumption to GCN DSL

Marks a variable as living in (0, 1): sets sympy positive=True plus an
inert unit_interval marker the steady-state solver reads to pick a logit
transform.

+++ gEconpy/solvers/shared.py
@@ -72,4 +72,4 @@ def o1_policy_function_adjoints(
 
 
 def pt_compute_selection_matrix(B, C, D, T):
-    return -pt.linalg.solve(C @ T + B, D.astype(T.dtype), assume_a="gen", check_finite=False)
+    return -pt.linalg.solve(C @ T + B, D, assume_a="gen", check_finite=False)
